backend/requirements.txt
django
djangorestframework
django-filter
gpxpy
gunicorn
dj-database-url
psycopg2-binary
django-cors-headers